requirements.txt
prompt-toolkit
python-dotenv
openai>=1.0.0
